+++ include/database/user_details.h
@@ -8,7 +8,7 @@ namespace balancedbanana {
 
         // Encapsulates all details required to create a user object
         struct user_details {
-            uint64_t id;
+            uint8_t id;
 
             std::string name;
